ubumtu 20.04 lts

+++ NLDFT_SDA_Slit_beta_v0.9.6_float/run.sh
@@ -13,7 +13,7 @@ if [ ! -d results ]; then
 fi
 
 #for w in "${items[@]}"; do
-for ((i=30;i<=1000;i+=2))
+for ((i=30;i<=1000;i+=1))
 do
 	w=`echo $i | awk '{printf "%4.2f", $1/100}'`
 	#echo $w
@@ -23,12 +23,18 @@ do
 	echo "Slit width = ${H} [nm]"
 	sed -i "s/XXX/${H}/g" parameters.txt
 	./nldft_sda_slit.exe
-	mv PP0_vs_Vgamma_data_vs.txt ${w}_data_vs.txt
-	mv PP0_vs_Vgamma_data_ls.txt ${w}_data_ls.txt
-	mv Pa_vs_Vgamma_data_vs.txt ${w}_data_vs.txt
-	mv Pa_vs_Vgamma_data_ls.txt ${w}_data_ls.txt
-	mv atm_vs_Vgamma_data_vs.txt ${w}_data_vs.txt
-	mv atm_vs_Vgamma_data_ls.txt ${w}_data_ls.txt
+	if [ -e PP0_vs_Vgamma_data_vs.txt ]; then
+	  mv PP0_vs_Vgamma_data_vs.txt ${w}_data_vs.txt
+	  mv PP0_vs_Vgamma_data_ls.txt ${w}_data_ls.txt
+	fi
+	if [ -e Pa_vs_Vgamma_data_vs.txt ]; then
+	  mv Pa_vs_Vgamma_data_vs.txt ${w}_data_vs.txt
+	  mv Pa_vs_Vgamma_data_ls.txt ${w}_data_ls.txt
+	fi
+	if [ -e atm_vs_Vgamma_data_vs.txt ]; then
+	  mv atm_vs_Vgamma_data_vs.txt ${w}_data_vs.txt
+	  mv atm_vs_Vgamma_data_ls.txt ${w}_data_ls.txt
+	fi
 	cp ${w}_data_vs.txt ./results/${w}_data_vs.txt
 	cp ${w}_data_ls.txt ./results/${w}_data_ls.txt
 	rm parameters.txt
